+++ azkaban-db/src/main/java/azkaban/db/DatabaseTransOperatorImpl.java
@@ -19,6 +19,8 @@ import java.sql.Connection;
 import java.sql.SQLException;
 import org.apache.commons.dbutils.QueryRunner;
 import org.apache.commons.dbutils.ResultSetHandler;
+import org.apache.commons.dbutils.handlers.ScalarHandler;
+import org.apache.log4j.Logger;
 
 
 /**
@@ -26,6 +28,7 @@ import org.apache.commons.dbutils.ResultSetHandler;
  */
 class DatabaseTransOperatorImpl implements DatabaseTransOperator {
 
+  private static final Logger logger = Logger.getLogger(DatabaseTransOperator.class);
   private final Connection conn;
   private final QueryRunner queryRunner;
 
@@ -33,6 +36,29 @@ class DatabaseTransOperatorImpl implements DatabaseTransOperator {
     this.conn = conn;
     this.queryRunner= queryRunner;
   }
+
+  /**
+   * The ID that was generated is maintained in Mysql server on a per-connection basis.
+   * This means that the value returned by the function to a given client is
+   * the first AUTO_INCREMENT value generated for most recent statement
+   *
+   * This value cannot be affected by other callers, even if they generate
+   * AUTO_INCREMENT values of their own.
+   * @return last insertion ID
+   *
+   */
+  @Override
+  public long getLastInsertId() {
+    // A default connection: autocommit = true.
+    long num = -1;
+    try {
+      num = ((Number) queryRunner.query(conn,"SELECT LAST_INSERT_ID();", new ScalarHandler<>(1))).longValue();
+    } catch (SQLException ex) {
+      logger.error("can not get last insertion ID", ex);
+    }
+    return num;
+  }
+
 
   @Override
   public <T> T query(String querySql, ResultSetHandler<T> resultHandler, Object... params) throws SQLException {
@@ -56,5 +82,9 @@ class DatabaseTransOperatorImpl implements DatabaseTransOperator {
     } finally {
       // Note: CAN NOT CLOSE CONNECTION HERE.
     }
+  }
+
+  public Connection getConnection() {
+    return conn;
   }
 }
